Add .clang-format file tailored for libnica
The format file is up to date for clang-format version 3.7 and has
comments suggesting changes to add for when clang-format version 3.9 is
available widely.

A few things that aren't really great that should change in future
versions:

1) string constants will not be broken up based on line length
2) comments will not need to be readjusted manually when their length is
greater than the column limit
3) include sorting by filename and type

Also there is a special comment pragma for doxygen style comments. This
can cause unfortunate comment alignment as they will never be split.
Care should be taken for ordering lines to have the lines that are over
the column limit come first and the rest of the comments will align
under them.