Ensure we are very explicit with our intention in using asprintf
It's currently ambiguous and unreliable what we truly mean with the
negated test for asprintf. Per the documentation, we should check that
the return was less than zero, indicating an issue with the asprintf
call itself.

This is also more helpful to tooling that would look at the return code
for branch prediction to more clearly state what the negative path would
actually be.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>