inifile: Do not abort when encountering internal errors
In all liklihood we've run out of memory or hit corruption, however it is
still not the place of a *library* to abort, it is the decision of the caller
alone.

Thus, we bail out of everything we're doing, and return a negative error code
indicating an internal error has happened.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>