Remove notion of NICA_INTERNAL
This is more how libraries like glib are implemented, by forcing the
use of a global header, and not being able to just pick a single header
to use.

This change makes nica more embeddable in other projects, without having
to maintain a diff against them. This has been seen currently in all
of the using projects: clr-boot-manager, linux-steam-integration, etc.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>