Use relative paths in the Makefile for all assets
For nica to be embedded in another project as a submodule or similar,
it would no longer be in the top source directory. To make this more
trivial, while allowing nica to continue to develop as a first class
library with it's own build system, we use relative pathing to enable
projects to add libnica to it's SUBDIRS in autotools.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>