Add support to integrate nica as a git submodule
For other projects to embed a copy of nica, they can now include the
entire nica tree as a git submodule in their tree. They should define
NICA_SUBMODULE to a truth value before doing so to disable the shared
library, valgrind + test suite, and installation of nica into the root.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>