Upstream the Nica Writer from clr-boot-manager (issue #12)
Effectively, this is a memory based write stream, designed to allow failure.
It is designed to allow multiple append/printf style invocations, and then
only testing for memory failures when this has all been done.

The style of the API usage largely follows that of the CbmMappedFile system
in clr-boot-manager, in attempting to avoid as many copies as possible, by
using a pointer to an anonymous stack struct, and associated open/close
methods. Due to the way that the memstream API works, we require both a
free *and* a close.

The close method is responsible for finalizing the buffer, and cleaning up
the associated file resource. This in turn ensures the null terminator is
placed into the final buffer. The free method is then hooked up into the
autofree system to ensure we never leak.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>