Upstream the Nica logger API from clr-boot-manager
The key differences between this logger and the cbm one, is correct
namespacing of the API, and defaulting to the WARNING log level.
Additionally we no longer emit a package name in the debug output as
it's very unnecessary.

Safety handling has been added to handle cases whereby the logger part
of libnica may be statically linked to to a final binary, thus destroying
the library constructors. In this case we fall back to a "maybe init"
approach and ensure we always have a valid log file, i.e. stderr fallback.

The file name and line number are now only displayed on errors, warnings,
fatal and debug messages, whereas the remainder will not emit these, to
make the output somewhat cleaner.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>